GC-GAN-main/GC-GAN-main/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
category-encoders==2.3.0
certifi==2021.10.8
cffi==1.15.0
cftime==1.6.2
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
ema-pytorch==0.0.9
entrypoints==0.4
executing==0.8.3
fonttools==4.29.1
googledrivedownloader==0.4
idna==3.3
importlib-resources==6.1.0
ipykernel==6.10.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.0
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.2
lxml==4.9.3
madgrad==1.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mne==1.5.1
mne-connectivity==0.5.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
netCDF4==1.6.4
networkx==2.6.3
nibabel==3.2.2
notebook==6.4.10
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.10.0
pooch==1.7.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
rdflib==6.1.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
statsmodels==0.13.2
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
# torch==1.8.0+cu111
# torch-ema==0.3
# torch-geometric==2.0.3
# torch-scatter==2.0.9
# torch-scatter==2.0.8
# torch-sparse==0.6.12
# torchaudio==0.8.0
# torchsummary==1.5.1
# torchvision==0.9.0+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
xarray==2023.1.0
yacs==0.1.8
zipp==3.7.0
